Added to documentation and improved variable names
Added meaning of 'direction' property to the documentation for pole_mount and comments (it's a bearing).

Replaced df_current_sheet variable with two variables that actually indicate which sheet is being referenced. (I delete the first one when it's no longer used - del keyword - so no extra memory is used.)

Added a link to the SpidaCalc documentation in the comments of xls-spida2csv-geodata.py so it's easier for people who read the code to figure out what the information it reads in represents.

A few other explanatory comments, ex: what WEP means, explaining why pole_spacing is originally assigned to a string that identifies one, (it's a placeholder, the identified WEP has the information that goes here).

Also renamed variable 'key' in a for loop to wireID, and created an intermediate variable called wepID to shorten a line of code and make it clear that the for loop is replacing the placeholder explained above.