Documentation for Spida converter
Create a documentation page for the Spida to csv converter under converters, fill in the broad overview from comments that were already in the python file, and additional details on interpreting the input data and output results. I explained the more unintuitive definitions, and added a link to a relevant part of the SpidaCalc documentation for easy reference.

Improved comments in xls-spida2csv-geodata.py